Feat/refactor start countdown

+++ Assets/Scripts/Data/MatchContext.cs
@@ -1,4 +1,4 @@
 public static class MatchContext
 {
-    public static int RoomId { get; set; }
+    public static long RoomId { get; set; }
 }
